+++ src/Form/Type/ServiceDeviceReferentSearchType.php
@@ -12,19 +12,20 @@ use App\Repository\Organization\PoleRepository;
 use App\Repository\Organization\ServiceRepository;
 use App\Repository\Organization\SubServiceRepository;
 use App\Repository\Organization\UserRepository;
-use App\Security\CurrentUserService;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Security\Core\Security;
 
 class ServiceDeviceReferentSearchType extends AbstractType
 {
-    private $currentUser;
+    /** @var User */
+    private $user;
 
-    public function __construct(CurrentUserService $currentUser)
+    public function __construct(Security $security)
     {
-        $this->currentUser = $currentUser;
+        $this->user = $security->getUser();
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options): void
@@ -35,7 +36,7 @@ class ServiceDeviceReferentSearchType extends AbstractType
         /** @var Service $service */
         $service = $attr['service'] ?? null;
 
-        if ($this->currentUser->hasRole('ROLE_SUPER_ADMIN')) {
+        if ($this->user->hasRole('ROLE_SUPER_ADMIN')) {
             $builder
                 ->add('poles', EntityType::class, [
                     'class' => Pole::class,
@@ -60,7 +61,7 @@ class ServiceDeviceReferentSearchType extends AbstractType
                     'choice_label' => 'name',
                     'multiple' => true,
                     'query_builder' => function (ServiceRepository $repo) use ($dataClass) {
-                        return $repo->getServicesOfUserQueryBuilder($this->currentUser, $dataClass);
+                        return $repo->getServicesOfUserQueryBuilder($this->user, $dataClass);
                     },
                     'attr' => [
                         'class' => 'multi-select w-min-150 w-max-200',
@@ -78,7 +79,7 @@ class ServiceDeviceReferentSearchType extends AbstractType
                     'choice_label' => 'name',
                     'multiple' => true,
                     'query_builder' => function (SubServiceRepository $repo) use ($service, $dataClass) {
-                        return $repo->getSubServicesOfUserQueryBuilder($this->currentUser, $service, $dataClass);
+                        return $repo->getSubServicesOfUserQueryBuilder($this->user, $service, $dataClass);
                     },
                     'attr' => [
                         'class' => 'multi-select w-min-160 w-max-200',
@@ -96,7 +97,7 @@ class ServiceDeviceReferentSearchType extends AbstractType
                     'choice_label' => 'name',
                     'multiple' => true,
                     'query_builder' => function (DeviceRepository $repo) use ($service, $dataClass) {
-                        return $repo->getDevicesOfUserQueryBuilder($this->currentUser, $service, $dataClass);
+                        return $repo->getDevicesOfUserQueryBuilder($this->user, $service, $dataClass);
                     },
                     'attr' => [
                         'class' => 'multi-select w-min-150 w-max-200',
@@ -114,7 +115,7 @@ class ServiceDeviceReferentSearchType extends AbstractType
                     'choice_label' => 'fullname',
                     'multiple' => true,
                     'query_builder' => function (UserRepository $repo) use ($service, $dataClass) {
-                        return $repo->getReferentsOfServicesQueryBuilder($this->currentUser, $service, $dataClass);
+                        return $repo->getReferentsOfServicesQueryBuilder($this->user, $service, $dataClass);
                     },
                     'attr' => [
                         'class' => 'multi-select w-min-150 w-max-220',