+++ src/Service/Event/TaskPaginator.php
@@ -2,26 +2,29 @@
 
 namespace App\Service\Event;
 
+use App\Entity\Organization\User;
 use App\Entity\Support\SupportGroup;
 use App\Form\Model\Event\TaskSearch;
 use App\Repository\Event\TaskRepository;
-use App\Security\CurrentUserService;
 use App\Service\Pagination;
 use Psr\Cache\CacheItemInterface;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Security;
 
 class TaskPaginator
 {
     private $pagination;
     private $taskRepo;
-    private $currentUserService;
 
-    public function __construct(Pagination $pagination, TaskRepository $taskRepo, CurrentUserService $currentUserService)
+    /** @var User */
+    private $user;
+
+    public function __construct(Pagination $pagination, TaskRepository $taskRepo, Security $security)
     {
         $this->pagination = $pagination;
         $this->taskRepo = $taskRepo;
-        $this->currentUserService = $currentUserService;
+        $this->user = $security->getUser();
     }
 
     /**
@@ -32,7 +35,7 @@ class TaskPaginator
         // Si filtre ou tri utilisé, n'utilise pas le cache.
         if (null === $supportGroup || $request->query->count() > 0) {
             return $this->pagination->paginate(
-                $this->taskRepo->findTasksQuery($search, $this->currentUserService, $supportGroup),
+                $this->taskRepo->findTasksQuery($search, $this->user, $supportGroup),
                 $request
             );
         }
@@ -44,7 +47,7 @@ class TaskPaginator
                 $item->expiresAfter(\DateInterval::createFromDateString('7 days'));
 
                 return $this->pagination->paginate(
-                    $this->taskRepo->findTasksQuery($search, $this->currentUserService, $supportGroup),
+                    $this->taskRepo->findTasksQuery($search, $this->user, $supportGroup),
                     $request
                 );
             }